HACK: Deploy only Fedora 37 packages
When deploying packages, build the RPM package for Fedora 37 and deploy
only that.

Problem
=======

The v0.4.0 release introduced Fedora 37 support. We had tested building
an RPM package for Fedora 37 and installing it successfully, so we
assumed that we had the green light for the release.

Turns out however that our CI runners were not building packages for
each of our supported Fedora release, but solely for Fedora 35. It so
happened that Fedora 35 and Fedora 36 had similar Python versions, but
the same did not apply to Fedora 37.

The end result is that users could not install Dangerzone on Fedora 37
through our official repo.

Remediation
===========

In order to help out Fedora 37 users, we plan to do the following:

1. Bump the Fedora patch level of our package to -2. This will help
   users install the new package, once we release it.
2. Change our CI pipeline to build an RPM package for Fedora 37 instead,
   and deploy that to PackageCloud.
3. Comment out all the other deployments to PackageCloud, since the rest
   of the distros are not affected.

IMPORTANT: This is a hotfix that should never be merged back in the main
branch.

Fixes #156